+++ test/spec/idx/remediate.ts
@@ -45,20 +45,16 @@ describe('idx/remediate', () => {
     expect(res).toEqual({ idxResponse });
   });
 
-  it('returns idxResponse, messages, terminal if the idxResponse is terminal', async () => {
+  it('returns idxResponse, terminal if the idxResponse is terminal', async () => {
     const { authClient } = testContext;
-    const messages = ['fake'];
     jest.spyOn(util, 'isTerminalResponse').mockReturnValue(true);
-    jest.spyOn(util, 'getMessagesFromResponse').mockReturnValue(messages);
     const idxResponse = { fake: true, actions: {} } as unknown as IdxResponse;
     const res = await remediate(authClient, idxResponse, {}, {});
     expect(res).toEqual({
       idxResponse,
       terminal: true,
-      messages
     });
     expect(util.isTerminalResponse).toHaveBeenCalledWith(idxResponse);
-    expect(util.getMessagesFromResponse).toHaveBeenCalledWith(idxResponse);
   });
 
   describe('actions', () => {
@@ -98,7 +94,7 @@ describe('idx/remediate', () => {
         });
 
         it('will handle exceptions', async () => {
-          const { authClient, remediator, errorResponse } = testContext;
+          const { authClient, errorResponse } = testContext;
           const error = new Error('my test error');
           const idxResponse = {
             neededToProceed: [{
@@ -111,7 +107,7 @@ describe('idx/remediate', () => {
           } as unknown as IdxResponse;
           const res = await remediate(authClient, idxResponse, { resend: true }, {});
           expect(res).toBe(errorResponse);
-          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, remediator);
+          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, {});
         });
       });
 
@@ -150,7 +146,7 @@ describe('idx/remediate', () => {
           expect(util.getRemediator).toHaveBeenNthCalledWith(2, responseFromAction.neededToProceed, {}, { actions: [] });
         });
         it('will handle exceptions', async () => {
-          const { authClient, remediator, errorResponse } = testContext;
+          const { authClient, errorResponse } = testContext;
           const error = new Error('my test error');
           const idxResponse = {
             neededToProceed: [{
@@ -161,9 +157,10 @@ describe('idx/remediate', () => {
             },
             rawIdxState: {}
           } as unknown as IdxResponse;
-          const res = await remediate(authClient, idxResponse, { resend: true }, { actions: ['some-action'] });
+          const options = { actions: ['some-action'] };
+          const res = await remediate(authClient, idxResponse, { resend: true }, options);
           expect(res).toEqual(errorResponse);
-          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, remediator);
+          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, options);
         });
       });
 
@@ -197,7 +194,7 @@ describe('idx/remediate', () => {
         });
         it('will handle exceptions', async () => {
           let { authClient, idxResponse } = testContext;
-          const { remediator, errorResponse } = testContext;
+          const { errorResponse } = testContext;
           const error = new Error('my test error');
           idxResponse = {
             ...idxResponse,
@@ -206,9 +203,10 @@ describe('idx/remediate', () => {
               name: 'some-remediation'
             }],
           } as unknown as IdxResponse;
-          const res = await remediate(authClient, idxResponse, {}, { actions: ['some-remediation'] });
+          const options = { actions: ['some-remediation'] };
+          const res = await remediate(authClient, idxResponse, {}, options);
           expect(res).toBe(errorResponse);
-          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, remediator);
+          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, options);
         });
       });
 
@@ -279,7 +277,7 @@ describe('idx/remediate', () => {
         });
         it('will handle exceptions', async () => {
           let { authClient, idxResponse } = testContext;
-          const { remediator, errorResponse } = testContext;
+          const { errorResponse } = testContext;
           const error = new Error('my test error');
           const actionFn = jest.fn().mockRejectedValue(error);
           idxResponse = {
@@ -295,9 +293,10 @@ describe('idx/remediate', () => {
             name: 'some-action',
             params: { foo: 'bar' }
           };
-          const res = await remediate(authClient, idxResponse, {}, { actions: [action] });
+          const options = { actions: [action] };
+          const res = await remediate(authClient, idxResponse, {}, options);
           expect(res).toBe(errorResponse);
-          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, remediator);
+          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, options);
           expect(actionFn).toHaveBeenCalledWith({ foo: 'bar' });
         });
       });
@@ -335,7 +334,7 @@ describe('idx/remediate', () => {
         });
         it('will handle exceptions', async () => {
           let { authClient, idxResponse } = testContext;
-          const { remediator, errorResponse } = testContext;
+          const { errorResponse } = testContext;
           const error = new Error('my test error');
           idxResponse = {
             ...idxResponse,
@@ -347,9 +346,10 @@ describe('idx/remediate', () => {
           const action = {
             name: 'some-remediation'
           };
-          const res = await remediate(authClient, idxResponse, {}, { actions: [action] });
+          const options = { actions: [action] };
+          const res = await remediate(authClient, idxResponse, {}, options);
           expect(res).toBe(errorResponse);
-          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, remediator);
+          expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, options);
         });
       });
 
@@ -437,9 +437,10 @@ describe('idx/remediate', () => {
             }]
           }],
         } as unknown as IdxResponse;
-        const res = await remediate(authClient, idxResponse, {}, { step: 'some-remediation' });
+        const options = { step: 'some-remediation' };
+        const res = await remediate(authClient, idxResponse, {}, options);
         expect(res).toBe(errorResponse);
-        expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error);
+        expect(util.handleIdxError).toHaveBeenCalledWith(authClient, error, options);
         expect(idxResponse.proceed).toHaveBeenCalledWith('some-remediation', {});
       });
     });
@@ -466,20 +467,6 @@ describe('idx/remediate', () => {
         idxResponse
       };
     });
-    // This behavior is being removed
-    // xit('if there are messages on the remediation, it returns early, without evaluation by the Remediator', async () => {
-    //   const { idxResponse, remediator } = testContext;
-    //   const messages = ['fake'];
-    //   jest.spyOn(util, 'getMessagesFromResponse').mockReturnValue(messages);
-    //   const res = await remediate(idxResponse, {}, {});
-    //   expect(res).toEqual({
-    //     idxResponse,
-    //     messages,
-    //     nextStep: {}
-    //   });
-    //   expect(util.getMessagesFromResponse).toHaveBeenCalledWith(idxResponse);
-    //   expect(remediator.canRemediate).not.toHaveBeenCalled();
-    // });
     it('if the Remediator cannot remediate, it returns early with nextStep information', async () => {
       const { authClient, idxResponse, remediator } = testContext;
       remediator.canRemediate.mockReturnValue(false);
@@ -525,7 +512,6 @@ describe('idx/remediate', () => {
             ...responseFromProceed,
             requestDidSucceed: true
           },
-          messages: ['hello'],
           nextStep: {}
         });
         expect(idxResponse.proceed).toHaveBeenCalledWith(name, data);
@@ -606,7 +592,6 @@ describe('idx/remediate', () => {
           requestDidSucceed: true,
         },
         terminal: true,
-        messages: ['hello'],
       });
     });
   });